Check TODO messages to be still valid
* According to index on GitHub
* Moved MultiValueDictionary (back) to AasxCsharLib
