+++ src/AasxCsharpLibrary/AdminShell.cs
@@ -2575,7 +2575,6 @@ namespace AdminShellNS
             void System.IDisposable.Dispose() { }
             public void GetData() { }
             // from HasDataSpecification
-            // TODO (Michael Hoffmeister, 1970-01-01): in V1.0, shall be a list of embeddedDataSpecification
             [XmlElement(ElementName = "embeddedDataSpecification")]
             [JsonIgnore]
             public EmbeddedDataSpecification embeddedDataSpecification = new EmbeddedDataSpecification();
@@ -3525,7 +3524,7 @@ namespace AdminShellNS
             public SemanticId semanticId = null;
 
             // this class
-            // TODO (Michael Hoffmeister, 1970-01-01): check, if Json has Qualifiers or not
+            // TODO (Michael Hoffmeister, 2020-08-01): check, if Json has Qualifiers or not
 
             [MetaModelName("Qualifier.type")]
             [TextSearchable]
