+++ src/AasxPackageExplorer/AasxPackageExplorer.csproj.user
@@ -3,4 +3,7 @@
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x64'">
     <StartArguments>-read-json options-debug.json</StartArguments>
   </PropertyGroup>
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|AnyCPU'">
+    <StartArguments>-read-json options-debug.json</StartArguments>
+  </PropertyGroup>
 </Project>
